backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pandas==2.3.2
numpy==2.3.2
scipy==1.16.1
statsmodels==0.14.5
openpyxl==3.1.5
matplotlib==3.10.5
seaborn==0.13.2
celery==5.5.3
redis==6.4.0
pydantic==2.11.7
python-dotenv==1.1.1
anthropic==0.60.0
reportlab==4.4.3
itsdangerous==2.2.0
